Move MOSS_working_example.pdf
Generated vignettes belong in inst/doc. This directory is not commonly
checked into version control, so we have to unignore it. This change
leads to a warning during R CMD check that can safely be ignored. Once
the package is published, I would suggest linking to the vignette on
CRAN.